bad brain: get first operation only if name is none
